+++ FIX/get_order.py
@@ -15,7 +15,7 @@ import sys
 import quickfix
 from Model.configuration import Configuration
 import configparser
-from orders import Orders
+from build_get import BuildGet
 
 config = configparser.ConfigParser()
 
@@ -26,14 +26,14 @@ def main():
         Configuration().build_config()
         settings = quickfix.SessionSettings('example.cfg', True)
 
-        orders_workshop = Orders()
+        build = BuildGet()
 
         storefactory = quickfix.FileStoreFactory(settings)
         logfactory = quickfix.FileLogFactory(settings)
-        initiator = quickfix.SSLSocketInitiator(orders_workshop, storefactory, settings, logfactory)
+        initiator = quickfix.SSLSocketInitiator(build, storefactory, settings, logfactory)
 
         initiator.start()
-        orders_workshop.run_get_order()
+        build.run_get_order()
 
     except (quickfix.ConfigError, quickfix.RuntimeError):
